Refatoração das camadas para padronização

+++ teart/src/main/java/com/duckers/teart/repositories/Interfaces/EnderecoUsuarioJPA.java
@@ -3,5 +3,5 @@ package com.duckers.teart.repositories.Interfaces;
 import com.duckers.teart.entities.EnderecoUsuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EnderecoUsuarioJPA extends JpaRepository<EnderecoUsuario, Integer> {
+public interface EnderecoUsuarioJPA extends JpaRepository<EnderecoUsuario, Long> {
 }
